[web] Fix thread creation from media
Summary:
This solves [ENG-5145](https://linear.app/comm/issue/ENG-5145/creating-a-sidebar-by-sending-an-image-doesnt-work-on-prod-web), a regression introduced in D7434.

The issue is that pending threads don't appear in the `ThreadStore`. Instead of trying to pull the `ThreadInfo` from the `ThreadStore`, we should drill it through from the callsites.

Test Plan: I confirmed that I'm again able to create a thread from media in my local environment (the bug no longer repros)

Reviewers: bartek, atul, tomek

Subscribers: