[keyserver] Replace extendCookieLifespan with updateCookie
Summary:
In the next diff this function will be used to replace a `bcrypt` hash with a `sha256` one. I want to do that in this function to avoid introducing any new SQL queries. `extendCookieLifespan` already runs very frequently, so it's a good place to make this change to ensure that we quickly migrate to `sha256` hashes.

Ahead of that change, this diff renames the function and updates it to take a full `Viewer` parameter.

Depends on D9546

Test Plan: Flow

Reviewers: atul, tomek

Reviewed By: tomek

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9548