[native] implement deleteUser native client method and expose it via CommRustModule
Summary: this is pretty similar to the updatePassword method. we just send the standard auth info (access token, user ID, device ID). since it's available in the comm secure store now, so we can probably refactor the JSI function in the future to not take any params. for now, though, i'm just following the pattern in updatePassword.

Test Plan: registered a new user and then called deleteUser from js and verified that the user's data was removed from my local ddb tables

Reviewers: bartek, kamil

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9555