[native] remove device ID getter and setter functions
Summary:
a device's ID is their content ed25519 signing public key. this gets created when the content olm account is created in the CryptoModule.

it seems like we can safely deprecate getDeviceID and setDeviceID and remove the callsite in sqlite-data-handler.js

Test Plan: successfully ran the app. tried logging in and out, registering and deleting an account. everything worked fine.

Reviewers: inka, kamil, marcin

Reviewed By: inka, kamil, marcin

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9556