[web] Support avatar uploads to blob service
Summary: This diff enables avatar uploads to blob service on web. Modified the hook to conditionally use function that uploads avatar to blob service instead of keyserver. This is very similar to what we currently do with multimedia messages.

Test Plan:
Tested on both user and thread avatars.
With the flag set to true, changed the avatar to an image and used TablePlus to verify MariaDB contents (this tests changes to D9168). Also, the avatar was then displayed correctly (which is more deeply tested in D9624)

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9170