[web] update AppearanceChangeModal buttons to use redux state
Summary:
Right now the updating the theme state is just localized to a `useState` in `AppearanceChangeModal`. Now that we have `globalThemeInfo` in our redux state we should update and use that instead.

Part of https://linear.app/comm/issue/ENG-4958/update-theme-based-on-the-active-theme-redux-state

Depends on D9553

Test Plan:
Please see the demo video below

{F827012}

Reviewers: atul, inka, rohan

Reviewed By: atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9554