[CI] run iOS Build CI on push to CI branch
Summary: I think this might let us avoid forking the repo to investigate CI issues. If the push event is to the CI branch, a tmate session will be set up.

Test Plan: will test this by pushing to CI branch on GitHub and seeing if iOS Build CI runs

Reviewers: ashoat, atul!

Subscribers: