[keyserver] Remove extraneous verifyClientSupported calls
Summary: In the past, we had calls to `validateInput` in each responder, which in turn calls `verifyClientSupported`. When the input did not need validation because there was no input, we would instead add direct calls to `verifyClientSupported`.

However, now `validateInput` is called implicitly before each responder. As such, `verifyClientSupported` is also called implicitly, and we don't need any explicit calls to it in responders.

Test Plan: Careful reading of the code. I think this change is relatively safe since I'm just removing a call that already occurs elsewhere

Reviewers: michal, marcin

Subscribers: