[RN72] Update React Native and React package versions
Summary:
This diff is the first step towards upgrading React Native.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Test Plan: The eventual test plan is to run the [iOS, Android] apps in [debug, release] on [simulator, physical device] and to play around with a bunch of the screens and make sure things generally work

Reviewers: tomek, bartek, atul

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9690