[RN72] Update Flow to 0.202.1
Summary:
The first thing I'm working on is fixing Flow errors with the new version of Flow, so I'll start by upgrading Flow.

The new version of Flow adds support for [enums](https://flow.org/en/docs/enums/), which I'm enabling here based on the lead of the React Native `.flowconfig`.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9690

Test Plan: The eventual test plan is to run the [iOS, Android] apps in [debug, release] on [simulator, physical device] and to play around with a bunch of the screens and make sure things generally work

Reviewers: tomek, bartek, atul, jon

Reviewed By: bartek, atul

Subscribers: wyilio

Differential Revision: https://phab.comm.dev/D9691