[CI] Get rid of KEYCHAIN_PASSWORD GitHub secret
Summary:
Based on [this doc](https://docs.github.com/en/actions/deployment/deploying-xcode-applications/installing-an-apple-certificate-on-macos-runners-for-xcode-development#creating-secrets-for-your-certificate-and-provisioning-profile), this value doesn't matter. We have it set to a random string via GitHub secrets, but we might as well just use some hardcoded string.

The benefit here is simplifying the set of secrets we need somebody to configure to get a fork of the Comm repo working with our CI workflows (so they can be debugged).

Test Plan: CI will test after landing

Reviewers: atul, varun

Subscribers: