[web] implement screen for testing Tunnelbroker
Summary:
This is a screen for testing Tunnelbroker utilities. It might be helpful for testing on production.

Styles/components are mostly copied from other places in the codebase which visually look similar. This is only for devs/staff and will be removed in the future.

Depends on D9616

Test Plan: See example [here](https://linear.app/comm/issue/ENG-5050/test-sending-messages-between-clients#comment-75a6e65d).

Reviewers: michal, bartek, tomek

Subscribers: ashoat, wyilio

Differential Revision: https://phab.comm.dev/D9617