[web] Display encrypted image avatars
Summary:
Added possibility to display encrypted image avatars on web.

Depends on D9167

Test Plan:
Requires some further diffs to be able to upload encrypted avatar to blob service.
Opened edit screen on web and selected new avatar. Confirmed it displays in the settings page as well as on the chat list and in the chat.

Reviewers: ginsu!, atul

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9624