[keyserver] Fetch user infos together with other data
Summary:
Instead of fetching user infos separately, we can fetch them with the rest of the infos.

Depends on D9656

Test Plan:
Generate some updates, reintroduce old code, and compare new with old result using `_.isEqual`.
For one update, generated during sending a friend request to an unknown user, the contents don't match. That seems to expose a race condition which was already present in our code (we're fetching the users a little earlier now) - going to investigate it further.

Reviewers: kamil, inka, bartek

Subscribers: ashoat, wyilio

Differential Revision: https://phab.comm.dev/D9657