[keyserver] implement socket and update code to interact with Tunnelbroker API
Summary:
1. Updates old code to new API
2. Implement Socket
3. Remove old types

There is some code duplication with what's in `lib/`. I'll try to refactor this later.

Looks like `RefreshKeyRequest` is defined not properly in Rust codebase, should not be a separate message but part of `MessageToDevice`. I'll fix that as followup - shouldn't affect this code.

Depends on D9596

Test plan:
1. Check if the keyserver can connect to Tunnelbroker
2. Check if the socket is properly closed when eg. Tunnelbroker is killed
3. Check if socket issues don't cause the keyserver to crash
4. Send `RefreshKeyRequest` and verify on keyserver and Identity logs that it is properly processed
4. Kill keyserver, send `RefreshKeyRequest`, launch keyserver and check if the message was delivered and processed.

Reviewers: michal, bartek, varun