[keyserver] Simplify updates merging
Summary:
The first step of moving merging updates into specs. The fact that we're doing the same thing in all the `if`s makes it possible to simplify the code significantly.

Depends on D9658

Test Plan: Tested in combination with the next diff. Generated some updates, reintroduced the old code, and compared merged updates.

Reviewers: kamil, inka, bartek

Subscribers: ashoat, wyilio

Differential Revision: https://phab.comm.dev/D9659