[native] Check isInvalidPinSource on native
Summary:
There was an inconsistency where I check `isInvalidPinSource` on `web`, but not on `native`. This diff corrects that. The following diff will cover unifying this logic into a shared helper so the checks are done in one place.

Resolves https://linear.app/comm/issue/ENG-5618/check-isinvalidpinsource-on-native-in-the-toolitp

Test Plan:
Confirmed that message pinning on `native` still works when expected:
- For text messages
- For multimedia messages
- In sidebars if the message is not the first message

Reviewers: atul, ginsu

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9634