[native] implement onPressSave keyserver button callback
Summary: This diff implements `onPressSave` keyserver button. This callback will create the payload necessary for this action and dispatch the add keyserver redux action we just introducted with the payload we just created. In a subsequent diff we will implement validation that will be used before we introduce the keyserver

Test Plan: Please see the demo video below

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9665