[lib] Avoid unnecessary Set construction in RadixTree.getAllMatchingExactly
Summary: Noticed this after landing the diff [here](https://phab.comm.dev/D9626#inline-60626). The underlying data is a `Set` anyways, so all this does it copy the `Set`. We don't return the original `Set` anyways, so there is no point to this.

Depends on D9646

Test Plan: Flow, careful code inspection

Reviewers: tomek, atul, inka, rohan

Subscribers: