[Flow202][lib][skip-ci] [8/23] Fix Flow issues in cookiesSelector
Summary:
The old version of Flow was treating the `{}` declaration as an "unsealed object", and as such was skipping typechecking it.

The new version of Flow doesn't do the "unsealed object" thing anymore, and forced me to annotate the declaration. That revealed a type error in `cookiesSelector`: the `cookies` in the returned object should be `?string`, but were typed as `string`.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9699

Test Plan: Confirm the Flow errors go away

Reviewers: inka

Reviewed By: inka

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9700