[Flow202][lib] [10/23] Fix Flow issues with untyped let declarations in entry-reducer
Summary:
The new version of Flow wants annotations on `let` expressions that cover later assignments.

Normally this just means adding an annotation, but in this case it's complicated because `updatedEntryInfos` is first treated as mutable and later as immutable.

To address this, I separated that variable out into two, allowing them to be typed differently.

NOTE: CI will fail on this diff. I considered the possibility of fixing Flow errors BEFORE upgrading Flow, but it wasn't possible... in some cases, the fixes to support the new version of Flow caused errors in the old version. I could have hidden these type errors with `$FlowFixMe` lines and then later revert those, but that seemed like too much busy work.

Depends on D9701

Test Plan: Confirm the Flow errors go away

Reviewers: atul

Reviewed By: atul

Subscribers: tomek, wyilio

Differential Revision: https://phab.comm.dev/D9702