[terraform] Use newer Amazon Linux 2 AMI for ECS instances
Summary:

Noticed this warning in ECS console:

Image here

This can be updated from the console for existing instances, but it's better to spawn new instances with the new image containing the latest ECS agent version.

Test Plan:

- Terraform plan finds correct image ID for the new image:

```
  # aws_launch_template.ecs_services will be updated in-place
  ~ resource "aws_launch_template" "ecs_services" {
      ~ default_version         = 1 -> (known after apply)
      ~ image_id                = "ami-0a2f86088203932e1" -> "ami-0b7778d86d8789cff"
```

This is the correct image ID for the new image from the AMI Catalog:

Image here.

Also tested on staging - new instances are launched with the new image and the agent version is correct.

Reviewers: kamil, michal, varun

Subscribers: